automations/seo_content_engine/requirements.txt
httpx>=0.27,<1.0
PyYAML>=6.0
Jinja2>=3.1
python-frontmatter>=1.1
